add oidc case to installation login test
